requirements.txt
async-generator==1.10
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
exceptiongroup==1.1.0
execnet==1.9.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
junit-xml==1.9
mypy-extensions==1.0.0
outcome==1.2.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pluggy==1.0.0
py==1.11.0
PySocks==1.7.1
pytest==7.2.1
pytest-html==3.2.0
pytest-metadata==2.0.4
pytest-xdist==3.1.0
PyYAML==6.0
requests==2.28.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
wsproto==1.2.0
